Move Task to model
