+++ app/src/debug/java/com/spanishdev/tasklistapp/domain/repository/DebugTaskRepositoryImpl.kt
@@ -1,7 +1,7 @@
 package com.spanishdev.tasklistapp.domain.repository
 
-import com.spanishdev.tasklistapp.domain.entities.Status
-import com.spanishdev.tasklistapp.domain.entities.Task
+import com.spanishdev.tasklistapp.domain.model.Status
+import com.spanishdev.tasklistapp.domain.model.Task
 
 class DebugTaskRepositoryImpl : TaskRepository {
