+++ app/src/main/java/com/spanishdev/tasklistapp/domain/usecase/UpdateTaskUseCase.kt
@@ -1,6 +1,6 @@
 package com.spanishdev.tasklistapp.domain.usecase
 
-import com.spanishdev.tasklistapp.domain.entities.Task
+import com.spanishdev.tasklistapp.domain.model.Task
 import com.spanishdev.tasklistapp.domain.repository.TaskRepository
 
 class UpdateTaskUseCase(private val repository: TaskRepository) {
